백엔드 기능 오류 잡음

+++ src/main/java/com/daelim/clover/likeList/controller/LikeListController.java
@@ -24,11 +24,17 @@ public class LikeListController {
     @ResponseBody
     public void insertLikeList(@RequestParam int boardId, Model model, HttpServletRequest request
     , int grade) throws Exception{
+
         HttpSession session = request.getSession();
         //로그인한 아이디
-        String userId = (String)session.getAttribute("sUserId");
-//        String userId = "qwe";
-        likeListService.insertLikeList(boardId,userId,grade);
+//        String userId = (String)session.getAttribute("sUserId");
+        String userId = "qwe";
+        if(userId !="" || userId != null){//유효성
+            likeListService.insertLikeList(boardId,userId,grade);
+        }else{
+            model.addAttribute("msg","로그인 해주세요");
+        }
+
     }
 
     //select
@@ -51,25 +57,38 @@ public class LikeListController {
     // 점수 수정
     @PutMapping("/updateLikeList")
     @ResponseBody
-    public void updateLikeList(@RequestParam int boardId,HttpServletRequest request ,@RequestParam int grade) throws Exception{
+    public void updateLikeList(@RequestParam int boardId,HttpServletRequest request ,Model model
+            ,@RequestParam int grade) throws Exception{
 
         HttpSession session = request.getSession();
         //로그인한 아이디
         String userId = (String)session.getAttribute("sUserId");
 //        String userId = "asd";
-        likeListService.updateLikeList(userId,boardId,grade);
+        if(userId !="" || userId != null){//유효성
+            likeListService.updateLikeList(userId,boardId,grade);
+        }else{
+            model.addAttribute("msg","로그인 해주세요");
+        }
+
     }
 
     // 점수 삭제
     @DeleteMapping("/deleteLikeList")
     @ResponseBody
-    public void deleteLikeList(@RequestParam int boardId,HttpServletRequest request) throws Exception{
+    public void deleteLikeList(@RequestParam int boardId,Model model
+            ,HttpServletRequest request) throws Exception{
 
         HttpSession session = request.getSession();
         //로그인한 아이디
-//        String userId = (String)session.getAttribute("sUserId");
-        String userId = "asd";
-        likeListService.deleteLikeList(boardId,userId);
+
+        String userId = (String)session.getAttribute("sUserId");
+//        String userId = "asd";
+        if(userId !="" || userId != null){//유효성
+            likeListService.deleteLikeList(boardId,userId);
+        }else{
+            model.addAttribute("msg","로그인 해주세요");
+        }
+
     }
 
 }
